+++ packages/core/src/agent/__tests__/tool-approval.test.ts
@@ -3,6 +3,7 @@ import z from 'zod';
 import { Mastra } from '../../mastra';
 import { InMemoryStore } from '../../storage';
 import { createTool } from '../../tools';
+import { createStep, createWorkflow } from '../../workflows';
 import { Agent } from '../agent';
 import { getOpenAIModel } from './mock-model';
 
@@ -264,7 +265,101 @@ export function toolApprovalAndSuspensionTests(version: 'v1' | 'v2') {
 
         expect(name).toBe('Dero Israel');
         expect(email).toBe('test@test.com');
-      }, 10000);
+      }, 15000);
+
+      it('should call findUserWorkflow with suspend and resume', async () => {
+        const findUserStep = createStep({
+          id: 'find-user-step',
+          description: 'This is a test step that returns the name and email',
+          inputSchema: z.object({
+            name: z.string(),
+          }),
+          suspendSchema: z.object({
+            message: z.string(),
+          }),
+          resumeSchema: z.object({
+            name: z.string(),
+          }),
+          outputSchema: z.object({
+            name: z.string(),
+            email: z.string(),
+          }),
+          execute: async ({ suspend, resumeData }) => {
+            if (!resumeData) {
+              return await suspend({ message: 'Please provide the name of the user' });
+            }
+
+            return {
+              name: resumeData?.name,
+              email: 'test@test.com',
+            };
+          },
+        });
+
+        const findUserWorkflow = createWorkflow({
+          id: 'find-user-workflow',
+          description: 'This is a test tool that returns the name and email',
+          inputSchema: z.object({
+            name: z.string(),
+          }),
+          outputSchema: z.object({
+            name: z.string(),
+            email: z.string(),
+          }),
+        })
+          .then(findUserStep)
+          .commit();
+
+        const userAgent = new Agent({
+          id: 'user-agent',
+          name: 'User Agent',
+          instructions: 'You are an agent that can get list of users using findUserWorkflow.',
+          model: openaiModel,
+          workflows: { findUserWorkflow },
+        });
+
+        const mastra = new Mastra({
+          agents: { userAgent },
+          logger: false,
+          storage: mockStorage,
+        });
+
+        const agentOne = mastra.getAgent('userAgent');
+
+        let toolCall;
+        const stream = await agentOne.stream('Find the user with name - Dero Israel');
+        const suspendData = {
+          suspendPayload: null,
+          suspendedToolName: '',
+        };
+        for await (const _chunk of stream.fullStream) {
+          if (_chunk.type === 'tool-call-suspended') {
+            suspendData.suspendPayload = _chunk.payload.suspendPayload;
+            suspendData.suspendedToolName = _chunk.payload.toolName;
+          }
+        }
+        if (suspendData.suspendPayload) {
+          const resumeStream = await agentOne.resumeStream({ name: 'Dero Israel' }, { runId: stream.runId });
+          for await (const _chunk of resumeStream.fullStream) {
+          }
+
+          const toolResults = await resumeStream.toolResults;
+
+          toolCall = toolResults?.find(
+            (result: any) => result.payload.toolName === 'workflow-findUserWorkflow',
+          )?.payload;
+
+          const name = toolCall?.result?.result?.name;
+          const email = toolCall?.result?.result?.email;
+
+          expect(name).toBe('Dero Israel');
+          expect(email).toBe('test@test.com');
+        }
+
+        expect(suspendData.suspendPayload).toBeDefined();
+        expect(suspendData.suspendedToolName).toBe('workflow-findUserWorkflow');
+        expect((suspendData.suspendPayload as any)?.message).toBe('Please provide the name of the user');
+      }, 15000);
     });
 
     describe.skipIf(version === 'v1')('persist model output stream state', () => {