+++ packages/core/src/agent/agent.ts
@@ -1685,11 +1685,11 @@ export class Agent<TAgentId extends string = string, TTools extends ToolsInput =
           subAgentResourceId: z.string().describe('The resource ID of the agent').optional(),
         });
 
-        const modelVersion = (await agent.getModel()).specificationVersion;
+        const modelVersion = (await agent.getModel({ requestContext })).specificationVersion;
 
         const toolObj = createTool({
           id: `agent-${agentName}`,
-          description: `Agent: ${agentName}`,
+          description: agent.getDescription() || `Agent: ${agentName}`,
           inputSchema: agentInputSchema,
           outputSchema: agentOutputSchema,
           mastra: this.#mastra,
@@ -1905,19 +1905,28 @@ export class Agent<TAgentId extends string = string, TTools extends ToolsInput =
                 resourceId,
               });
 
-              const run = await workflow.createRun();
+              const run = await workflow.createRun({ runId });
 
               const { initialState, inputData: workflowInputData } = inputData;
+              const { resumeData, suspend } = context?.agent ?? {};
 
               let result: WorkflowResult<any, any, any, any> | undefined = undefined;
 
               if (methodType === 'generate' || methodType === 'generateLegacy') {
-                result = await run.start({
-                  inputData: workflowInputData,
-                  requestContext,
-                  tracingContext: context?.tracingContext,
-                  ...(initialState && { initialState }),
-                });
+                if (resumeData) {
+                  result = await run.resume({
+                    resumeData,
+                    requestContext,
+                    tracingContext: context?.tracingContext,
+                  });
+                } else {
+                  result = await run.start({
+                    inputData: workflowInputData,
+                    requestContext,
+                    tracingContext: context?.tracingContext,
+                    ...(initialState && { initialState }),
+                  });
+                }
               } else if (methodType === 'streamLegacy') {
                 const streamResult = run.streamLegacy({
                   inputData: workflowInputData,
@@ -1935,12 +1944,18 @@ export class Agent<TAgentId extends string = string, TTools extends ToolsInput =
 
                 result = await streamResult.getWorkflowState();
               } else if (methodType === 'stream') {
-                const streamResult = run.stream({
-                  inputData: workflowInputData,
-                  requestContext,
-                  tracingContext: context?.tracingContext,
-                  ...(initialState && { initialState }),
-                });
+                const streamResult = resumeData
+                  ? run.resumeStream({
+                      resumeData,
+                      requestContext,
+                      tracingContext: context?.tracingContext,
+                    })
+                  : run.stream({
+                      inputData: workflowInputData,
+                      requestContext,
+                      tracingContext: context?.tracingContext,
+                      ...(initialState && { initialState }),
+                    });
 
                 if (context?.writer) {
                   await streamResult.fullStream.pipeTo(context.writer);
@@ -1959,10 +1974,13 @@ export class Agent<TAgentId extends string = string, TTools extends ToolsInput =
                   runId: run.runId,
                 };
               } else if (result?.status === 'suspended') {
-                return {
-                  error: `Workflow ended with status: "suspended". This is not currently handled in the basic agent workflow tool transformation. To achieve this you'll need to write your own tool that uses a workflow internally.`,
-                  runId: run.runId,
-                };
+                const suspendedStep = result?.suspended?.[0]?.[0]!;
+                const suspendPayload = result?.steps?.[suspendedStep]?.suspendPayload;
+
+                if (suspendPayload?.__workflow_meta) {
+                  delete suspendPayload.__workflow_meta;
+                }
+                return suspend?.(suspendPayload);
               } else {
                 // This is to satisfy the execute fn's return value for typescript
                 return {
@@ -2478,6 +2496,8 @@ export class Agent<TAgentId extends string = string, TTools extends ToolsInput =
       input: options.messages,
       attributes: {
         agentId: this.id,
+        agentName: this.name,
+        conversationId: threadFromArgs?.id,
         instructions: this.#convertInstructionsToString(instructions),
       },
       metadata: {
@@ -2519,6 +2539,8 @@ export class Agent<TAgentId extends string = string, TTools extends ToolsInput =
       getMemoryMessages: this.getMemoryMessages.bind(this),
       runInputProcessors: this.__runInputProcessors.bind(this),
       executeOnFinish: this.#executeOnFinish.bind(this),
+      inputProcessors: async ({ requestContext }: { requestContext: RequestContext }) =>
+        this.listResolvedInputProcessors(requestContext),
       outputProcessors: async ({ requestContext }: { requestContext: RequestContext }) =>
         this.listResolvedOutputProcessors(requestContext),
       llm,